Removed passthrough argument -- parser should now determine when to passthrough a parameter based on the function signature.
Parser will attempt to convert object types when the provided values do not match up exactly with the function signature. E.g. sampling a discrete distribution returns integers, but the function signature requirea a double.

ListOfInputs is no longer necessary, as any collection can be used for this purpose. Similar to the passthrough check, determining how to use a collection depends on the function signature.